- reverted to using uint16_t for line sensor min/max

+++ software/sensors-and-motors/lib-driver-stm32l496ze/src/global_data.c
@@ -35,8 +35,8 @@ volatile uint8_t g_imuTxBuffer[IMU_SPI_TX_BUFFER_LENGTH];
 
 volatile float g_lineDisplacementFromCenterlineInMeters;
 volatile uint16_t g_lineSensorValues[LINE_SENSOR_COUNT];
-volatile float g_lineSensorMinValues[LINE_SENSOR_COUNT];
-volatile float g_lineSensorMaxValues[LINE_SENSOR_COUNT];
+volatile uint16_t g_lineSensorMinValues[LINE_SENSOR_COUNT];
+volatile uint16_t g_lineSensorMaxValues[LINE_SENSOR_COUNT];
 
 /*******************************************************************************
  * spi                                                                         *
